refactor: added useCallback and made style changes
